requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
asgiref=3.2.3=py_0
boost-cpp=1.70.0=h75728bb_2
bzip2=1.0.8=h0b31af3_2
ca-certificates=2019.11.28=hecc5488_0
cairo=1.16.0=he1c11cd_1002
certifi=2019.11.28=py38_0
cfitsio=3.470=h84d2f63_2
curl=7.68.0=h8754def_0
django=3.0.3=py38_0
django-widget-tweaks=1.4.5=py_0
expat=2.2.9=h4a8c4bd_2
fontconfig=2.13.1=h6b1039f_1001
freetype=2.9.1=hb4e5f40_0
freexl=1.0.5=h1de35cc_1002
gdal=2.4.3=py38h1304821_8
geojson=2.5.0=py_0
geos=3.7.2=h6de7cb9_2
geotiff=1.5.1=h2bcb450_3
gettext=0.19.8.1=h46ab8bc_1002
giflib=5.1.7=h01d97ff_1
glib=2.58.3=py38h577aef8_1002
hdf4=4.2.13=h84186c3_1003
hdf5=1.10.5=nompi_h3e39495_1104
icu=64.2=h6de7cb9_1
jpeg=9c=h1de35cc_1001
json-c=0.13.1=h1de35cc_1001
kealib=1.4.10=h6659575_1005
krb5=1.16.4=h1752a42_0
libblas=3.8.0=14_openblas
libcblas=3.8.0=14_openblas
libcurl=7.68.0=h709d2b2_0
libcxx=9.0.1=1
libcxxabi=4.0.1=hcfea43d_1
libdap4=3.20.4=habf5908_0
libedit=3.1.20181209=hb402a30_0
libffi=3.2.1=h475c297_4
libgdal=2.4.3=h48e9a63_8
libgfortran=4.0.0=2
libiconv=1.15=h01d97ff_1005
libkml=1.3.0=hed7d534_1010
liblapack=3.8.0=14_openblas
libnetcdf=4.7.1=nompi_hec86efb_102
libopenblas=0.3.7=h3d69b6c_7
libpng=1.6.37=ha441bb4_0
libpq=11.5=h31a01ba_2
libspatialite=4.3.0a=hd0a3780_1030
libssh2=1.8.2=hcdc9a53_2
libtiff=4.1.0=ha78913b_3
libxml2=2.9.10=h53d96d6_0
llvm-openmp=9.0.1=h28b9765_2
lz4-c=1.8.3=h6de7cb9_1001
ncurses=6.1=h0a44026_1
numpy=1.18.1=py38hde6bac1_0
olefile=0.46=py_0
openjpeg=2.3.1=hcdae239_3
openssl=1.1.1d=h0b31af3_0
pandas=1.0.1=py38h4f17bb1_0
pcre=8.43=h4a8c4bd_0
pillow=7.0.0=py38h918e99a_0
pip=20.0.2=py38_1
pixman=0.38.0=h01d97ff_1003
poppler=0.67.0=h16886b5_8
poppler-data=0.4.9=1
postgresql=11.5=h26bc10f_2
proj4=6.1.1=hca663eb_1
psycopg2=2.8.4=py38hafa8578_0
python=3.8.1=hf2284b6_2
python-dateutil=2.8.1=py_0
pytz=2019.3=py_0
readline=8.0=hcfe32e1_0
setuptools=45.1.0=py38_0
six=1.14.0=py38_0
sqlite=3.30.1=ha441bb4_0
sqlparse=0.3.0=py_0
tk=8.6.10=hbbe82c9_0
tzcode=2019a=h01d97ff_1002
wheel=0.34.2=py38_0
xerces-c=3.2.2=h8f8adb3_1004
xz=5.2.4=h1de35cc_4
zlib=1.2.11=h1de35cc_3
zstd=1.4.4=he7fca8b_1
